+++ raytracer/Camera.h
@@ -8,18 +8,18 @@ public:
     static Camera& create_on_device(
         Point3 const& look_from,
         Point3 const& look_at,
-        double vertical_fov,
+        float vertical_fov,
         size_t image_width, size_t image_height);
 
     __device__ Camera(
         Point3 look_from,
         Point3 look_at,
         Vec3 view_up,
-        double vertical_fov,
+        float vertical_fov,
         size_t image_width,
         size_t image_height);
 
-    __device__ Ray get_ray(double u, double v) const
+    __device__ Ray get_ray(float u, float v) const
     {
         return { origin, lower_left_corner + u * horizontal + v * vertical - origin };
     }
