+++ raytracer/material/Dielectric.h
@@ -17,7 +17,7 @@ public:
         Ray const& r_in, HitRecord const& rec, Color& attenuation, Ray& scattered) const override;
 
 private:
-    static __device__ Vec3 refract_vector(Vec3 const& uv, Vec3 const& n, double etai_over_etat);
+    static __device__ Vec3 refract_vector(Vec3 const& uv, Vec3 const& n, float etai_over_etat);
     static __device__ float reflectance(float cosine, float ref_idx);
 
     DeviceRNG& m_rng;
